Increase ping timeout, add option & desired capability for ping timeout
Low ping timeout prevented driver from successfully connecting
to the app while creating new session due to bad network,
resulting in tests failure.
Ping timeout can be set both as cli option or desired capability.
Mostly it would be set as cli option, as it is specific to environment,
but it could be overridden by capability if needed.